% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DendroPlot.R
\name{DendroPlot}
\alias{DendroPlot}
\title{Plot dendrogram}
\usage{
DendroPlot(Cluster.object, No.Cluster = NULL, CH = FALSE, plain = FALSE, ...)
}
\arguments{
\item{Cluster.object}{The object return from function \code{LongDataCluster}}

\item{No.Cluster}{User specified number of clusters}

\item{CH}{If \code{TRUE}, use optimal number of clusters determined by CH index to cut dendrogram. The default is to use the \eqn{Gap_b} metric}

\item{plain}{If \code{TRUE}, the dendrogram only, without color coding the clustering results}

\item{...}{Additional arguments for plot}
}
\description{
Based on output from \code{LongDataCluster}, yield the corresponding dendropgram which demonstrates the merging process
}
\examples{
output = LongDataCluster(Longdat$Dat$obs,
Longdat$Dat[,paste("y", seq(5), sep = "_")],
Longdat$Dat$id)
DendroPlot(output)

}
