% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalCluster.R
\name{FinalCluster}
\alias{FinalCluster}
\title{Ancillary function for parallel computing (not for call)}
\usage{
FinalCluster(pure.leaf, stop = 1, weight.func, DistMetric, y.dim)
}
\arguments{
\item{pure.leaf}{Output from LongDataCluster.core in each server}

\item{stop}{When iteratively applying parallel idea, stop when a certain number of clusters are left}

\item{weight.func}{weight function}

\item{DistMetric}{Weighted or Unweighted distance metric}
}
\value{
The same output as LongDataCluster.core
}
\description{
Based on outputs from each server/slave, merge them and keep hierarchical algorithm until only one group left
}
\author{
Junyi Zhou \email{junyzhou@iu.edu}
}
