% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Clustering longitudinal data (not for call)}
\usage{
test(
  x,
  Y,
  id,
  functional = "bs",
  preprocess = TRUE,
  weight.func = "standardize",
  parallel = FALSE,
  stop = 20,
  ...
)
}
\arguments{
\item{x}{A vector in long format, occassions or time of observation times.}

\item{Y}{A matrix, if multiple outcomes exist; or a (column) vector, for single outcome case}

\item{id}{A vector with same length of x, represents the corresponding subject id of each observation}

\item{functional}{A string from c("bs", "ns"), indicating b-splines and natural splines}

\item{preprocess}{boolean, whether data pre-processing procedure should be applied. Default is TRUE to handle subjects with observations less than number of parameters. If set to FALSE, those subjects will be excluded.}

\item{weight.func}{A string from c("standardize", "softmax"), a method to handle weights across multiple outcomes. Default is "standardize", but "softmax" is recommended for cases with a lot noise (indistinguishable) outcomes}

\item{parallel}{boolean, indicating whether parallel computing should be implemented. Default is FALSE}

\item{stop}{An integer, only required in parallel computing. It indicates when to stop hierarchical clustering process in each slave/serve}

\item{...}{Additional arguments for the functional bases. The default is cubic B-spline with 3 interval knots}
}
\value{
A list object containing the hierarchical clustering results, and some ancillary outputs for parallel computing. The optimal number of clusters will not be determined by this function.
}
\description{
Clustering longitudinal data, expecially tailored for those observational longitudinal data with sparse and irregular observations. The output could be a vector, i.e., at each occasion, more than one measure are observed for each subject.
}
