% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanPlot.R
\name{MeanPlot}
\alias{MeanPlot}
\title{Plot estimated mean trajectories for each detected cluster}
\usage{
MeanPlot(Cluster.object, No.Cluster = NULL, CH = FALSE)
}
\arguments{
\item{Cluster.object}{The object return from function LongDataCluster}

\item{No.Cluster}{User specified number of clusters}

\item{CH}{If TRUE, iuse optimal number of clusters determined by CH index to cut dendrogram. The default is to use Gapb}
}
\value{
A figure yield by plotly, which mean trajectories in different colors.
}
\description{
Based on output from LongDataCluster, yield the corresponding mean curves
}
\examples{
\dontrun{
require(doMC)
registerDoMC(cores = 6)
output = LongDataCluster(Longdat2$Dat$obs,
Longdat2$Dat[,paste("y", seq(5), sep = "_")],
Longdat2$Dat$id, parallel = T, dropout = 15, part.size = 200)
MeanPlot(output)
}
}
